+++ Locale/esMX.lua
@@ -7,7 +7,7 @@
 --]]
 
 if (((TacoTipConfig and TacoTipConfig.locale_override) or GetLocale()) ~= "esMX") then return end
-TACOTIP_LOCALE = {
+_G.TACOTIP_LOCALE = {
 ["Player"] = "Jugador",
 ["Pet"] = "Mascota",
 ["Target"] = "Objetivo",