+++ Locale/ruRU.lua
@@ -9,7 +9,7 @@
 --]]
 
 if (((TacoTipConfig and TacoTipConfig.locale_override) or GetLocale()) ~= "ruRU") then return end
-TACOTIP_LOCALE = {
+_G.TACOTIP_LOCALE = {
 ["Player"] = "Игрок",
 ["Pet"] = "Питомец",
 ["Target"] = "Цель",